Category GET and Product POST
Finished the Category dropdown populating as well as successfully inserting a new product from the front end. Types are of all string, had to cast in the backend, which is correct.

Excess code for the fetch's regarding GET, need to create one method that takes in a URL endpoint to be more concise. other than that, Inserting functionality for product is complete. React is a bitch to understand