Cohesiveness and Dropdown Improvements
Backend:

Increased cohesion and independence of handler func's by giving a new handlerfunc for each request instead of using if/else statements within one func. Changed handler url for populating the dropdown for categories based on a dynamic vendor name and simplified querying for the categories by natural joining the vendor and category table. Nice K.

Frontend:

Changes to structure of populating the category dropdown, specifically in the returned html and within the fetchCategories func. The value of the vandorName required to send to the backend was changed on line 192 from an object call of an int value to the selectedVendor value within the value on like 201. Also changing value=vendorId to value = vendorName was required.

The fetchCategories func now takes in a properly named parameter of a string value that is selected from the vendor dropdown, and fetches the associated categories before populating them in the categories dropdown.

TODO:

If a vendor has no associated categories and it is selected, the app will crash. So when we set a new vendor, it must require one or more new categories to be associated with.

- Finish Inserting Product
- Begin Inserting Category and Vendor
- Deleting Product
- Modifying Product
- Displaying Products