basic validation, basic banners, began basic auth
Finished basic authentication for inserting new product with rudimentary banners.

Restructured Go-Server files for readability and modularity.

Began basic auth for inserting, it's a pain in the ass though.

Added react-toastify lib for response banners.

Added .json files to .gitignore as there is a package that is incompatible between mac/windows that I really cba to keep deleting