Reformatted Authentication
Overhauled authentication functions to take users username and pass, which are then used in a stored procedure in the database to compare to a hashed pass.

Stored Procedure: GetHashyPassy:

Allows the front and backend to take username and password provided. Encrpyts by salting the password then hashing. Then compares to the HashPass stored in the database. If equal, returns username tuple that is then used to determine success or failure. If a tuple is returned, success. If procedure returns no tuples, failure.